re_entry, rm .vs, images output
