+++ Lab2/WpfAppEmotions/WpfAppEmotions/AssemblyInfo.cs
