--- Lab2_v1/Lib1/EmotionFerPlusLib.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Title>EmotionFerPlusLib</Title>
-  </PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.ML.OnnxRuntime" Version="1.4.0" />
-		<PackageReference Include="SixLabors.ImageSharp" Version="1.0.1" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<EmbeddedResource Include="emotion-ferplus-7.onnx" />
-	</ItemGroup>
-
-</Project>
