--- Lab2_v1/WpfApp1/WpfApp1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32922.545
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApp1", "WpfApp1\WpfApp1.csproj", "{6F69BEB6-232E-4CBA-879E-49E123CFCFB1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6F69BEB6-232E-4CBA-879E-49E123CFCFB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F69BEB6-232E-4CBA-879E-49E123CFCFB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F69BEB6-232E-4CBA-879E-49E123CFCFB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F69BEB6-232E-4CBA-879E-49E123CFCFB1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B1FAC6CF-7F2D-4597-BB4E-266A665AD2C3}
-	EndGlobalSection
-EndGlobal
