--- Lab2_v1/WpfApp1/WpfApp1/EmotionConverter.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace WpfApp1
-{
-    internal class EmotionConverter
-    {
-    }
-}
