--- Lab2_v1/WpfApp1/WpfApp1/MainWindow.xaml
@@ -1,68 +1,0 @@
-﻿<Window x:Class="WpfApp1.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfApp1"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="1*"/>
-            <RowDefinition Height="2*"/>
-            <RowDefinition Height="1*"/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="1*"/>
-            <ColumnDefinition Width="1*"/>
-            <ColumnDefinition Width="1*"/>
-        </Grid.ColumnDefinitions>
-
-
-
-        <TextBlock  Grid.Row = "1" Grid.Column = "0" VerticalAlignment = "Top" Height="47">
-                Выберите эмоцию для <LineBreak/>сортировки списка картинок:
-        </TextBlock>
-        
-        <ComboBox x:Name="Emotion_order" Height="23" VerticalAlignment="Top" Grid.Column="0" Grid.Row="1" 
-                SelectedIndex = "{Binding selected_index}" Grid.ColumnSpan="2" Margin="0,47,266,0">
-            <ComboBoxItem Content = "neutral"/>
-            <ComboBoxItem Content = "happiness"/>
-            <ComboBoxItem Content = "surprise"/>
-            <ComboBoxItem Content = "sadness"/>
-            <ComboBoxItem Content = "anger"/>
-            <ComboBoxItem Content = "disgust"/>
-            <ComboBoxItem Content = "fear"/>
-            <ComboBoxItem Content = "contempt"/>
-        </ComboBox>
-
-
-
-        <Button Click="Open_Images_Click" Width="120" Grid.Column="0" Grid.Row="0" 
-                Height="40">Open images</Button>
-        <Button Click="Start_Calc_Click" Width="120" Grid.Column="1" Grid.Row="0" 
-                Height="40">Start calculations</Button>
-        <Button Click="Stop_Calc_Click" Width="120" Grid.Column="2" Grid.Row="0" 
-                Height="40">Cancel calculations</Button>
-
-        <ProgressBar Minimum="0" Maximum="100" Name="pbStatus" Value="{Binding MyCollection_sort.progress}" 
-                      Grid.Column="2" Grid.Row="2"/>
-        <TextBlock Text="{Binding ElementName = pbStatus, Path=Value, 
-                StringFormat={}{0:0.0}%}" HorizontalAlignment="Center" 
-                   VerticalAlignment="Center" Grid.Column="1" Grid.Row="1"/>
-
-        <ListBox ItemsSource="{Binding MyCollection_sort.MyList}"  Grid.Column="2" Grid.Row="1">
-            <ListBox.ItemTemplate>
-                <DataTemplate>
-                    <StackPanel>
-                        <TextBlock Text="{Binding image_name}"/>
-                        <ListBox ItemsSource="{Binding emotions}"/>
-                    </StackPanel>
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>
-
-    </Grid>
-</Window>
