--- Lab2_v1/WpfApp1/WpfApp1/WpfApp1.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
-    <Nullable>enable</Nullable>
-    <UseWPF>true</UseWPF>
-  </PropertyGroup>
-
-
-	<ItemGroup>
-		<PackageReference Include="microsoft.ml.onnxruntime" Version="1.12.1" />
-		<PackageReference Include="sixlabors.imagesharp" Version="2.1.3" />
-		<PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="EmotionFerPlusLib" Version="1.0.0" />
-	</ItemGroup>
-</Project>
